+++ branch-alpha/CHANGELOG.md
@@ -5,12 +5,13 @@
 - jslint - Audit token-property '.free'.
 - jslint-ecma - Expand ES2015-feature-support for es-module-export-statement.
 - jslint - Relax warning expected_line_break_a_b for ternary-operator inside template-literal.
-- jslint - Add ability to auto-fix whitespace.
+- jslint - Expand jslint_autofix to indentation expected_a_at_b_c and line-break expected_line_break_a_b, and to binary-operator-end-of-line.
 - jslint - Add html and css linting back into jslint.
 - jslint - Add new warning requiring paren around plus-separated concatenations.
 - jslint - Try to improve parser to be able to parse jquery.js without stopping.
 
 # v2026.9.29
+- jslint - Add command jslint_autofix to auto-fix whitespace-warnings expected_space_a_b and unexpected_space_a_b.
 - jslint - Add beta-warning to jslint_phase5_whitage(), requiring binary-operators at end-of-line, except dot-operators '.', '?.'.
 - jslint-ecma - Add ES2018-feature Asynchronous Iteration - for await...of.
 - jslint - Move 242 line-leading && or || to line end.